+++ projects/plugins/crm/src/js/components/automations-admin/index.tsx
@@ -5,7 +5,7 @@ import { Link, Routes, Route } from 'react-router-dom';
 import AdminPage from '../admin-page';
 import { RedirectHome } from './components/redirect-home';
 import { WorkflowRow } from './components/workflow-row';
-import { workflowOne, workflowTwo } from './test/test-data';
+import { workflowOne, workflowTwo } from './test/util/data';
 
 export const AutomationsAdmin = () => {
 	return (
