+++ projects/plugins/crm/src/js/components/automations-admin/test/util/data.ts
